Remove check for pth file
This error is already generated when the file is read. The hard-
coded path used by the deleted check is not necessarily the
source code location. The deleted check ends up requiring the pth
file in two locations - one specified by the hard-coded path and
the other used by the code that actually reads the file.